+++ src/Okta.Sdk.IntegrationTests/PoliciesScenarios.cs
@@ -46,6 +46,36 @@ namespace Okta.Sdk.IntegrationTests
         }
 
         [Fact]
+        public async Task CreateProfileEnrollmentPolicy()
+        {
+            var client = TestClient.Create();
+
+            var policy = new Policy()
+            {
+                // Name has a maximum of 50 chars
+                Name = $"dotnet-sdk: ProfileEnrollmentPolicy {Guid.NewGuid()}".Substring(0, 50),
+                Type = PolicyType.ProfileEnrollment,
+                Status = "ACTIVE",
+                Description = "The default policy applies in all situations if no other policy applies.",
+            };
+
+            var createdPolicy = await client.Policies.CreatePolicyAsync(policy);
+
+            try
+            {
+                createdPolicy.Should().NotBeNull();
+                createdPolicy.Name.Should().Be(policy.Name);
+                createdPolicy.Type.Should().Be(PolicyType.ProfileEnrollment);
+                createdPolicy.Status.Should().Be("ACTIVE");
+                createdPolicy.Description.Should().Be(policy.Description);
+            }
+            finally
+            {
+                await client.Policies.DeletePolicyAsync(createdPolicy.Id);
+            }
+        }
+
+        [Fact]
         public async Task GetPolicy()
         {
             var client = TestClient.Create();
@@ -428,6 +458,146 @@ namespace Okta.Sdk.IntegrationTests
             finally
             {
                 await client.Policies.DeletePolicyRuleAsync(createdPolicy.Id, createdPolicyRule.Id);
+                await client.Policies.DeletePolicyAsync(createdPolicy.Id);
+            }
+        }
+
+        [Fact]
+        public async Task CreateAccessPolicyPolicyRule()
+        {
+            var client = TestClient.Create();
+            var guid = Guid.NewGuid();
+
+            var createdApp = await client.Applications.CreateApplicationAsync(new CreateOpenIdConnectApplication
+            {
+                Label = $"dotnet-sdk: AddOpenIdConnectApp {guid}",
+                ClientUri = "https://example.com/client",
+                LogoUri = "https://example.com/assets/images/logo-new.png",
+                ResponseTypes = new List<OAuthResponseType>
+                {
+                    OAuthResponseType.Token,
+                    OAuthResponseType.IdToken,
+                    OAuthResponseType.Code,
+                },
+                RedirectUris = new List<string>
+                {
+                    "https://example.com/oauth2/callback",
+                    "myapp://callback",
+                },
+                PostLogoutRedirectUris = new List<string>
+                {
+                    "https://example.com/postlogout",
+                    "myapp://postlogoutcallback",
+                },
+                GrantTypes = new List<OAuthGrantType>
+                {
+                    OAuthGrantType.Implicit,
+                    OAuthGrantType.AuthorizationCode,
+                },
+                ApplicationType = OpenIdConnectApplicationType.Web,
+            });
+
+            var accessPolicyId = createdApp.GetAccessPolicyId();
+
+            var accessPolicyRuleOptions = new AccessPolicyRule
+            {
+                Name = $"dotnet-sdk: CreateAccessPolicyRule {guid}".Substring(0, 50),
+                Type = PolicyType.AccessPolicy.ToString(),
+                Actions = new AccessPolicyRuleActions
+                {
+                      AppSignOn = new AccessPolicyRuleApplicationSignOn
+                      {
+                          Access = "DENY",
+                          VerificationMethod = new VerificationMethod
+                          {
+                              Type = "ASSURANCE",
+                              FactorMode = "1FA",
+                              ReauthenticateIn = "PT43800H",
+                          },
+                      },
+                },
+            };
+
+            var createdPolicyRule =
+                await client.Policies.AddPolicyRuleAsync<AccessPolicyRule>(accessPolicyRuleOptions, accessPolicyId);
+            try
+            {
+                createdPolicyRule.Should().NotBeNull();
+                createdPolicyRule.Name.Should().Be(accessPolicyRuleOptions.Name);
+                createdPolicyRule.Actions.Should().NotBeNull();
+                createdPolicyRule.Actions.AppSignOn.Access.Should().Be("DENY");
+                createdPolicyRule.Actions.AppSignOn.VerificationMethod.Type.Should().Be("ASSURANCE");
+                createdPolicyRule.Actions.AppSignOn.VerificationMethod.FactorMode.Should().Be("1FA");
+                createdPolicyRule.Actions.AppSignOn.VerificationMethod.ReauthenticateIn.Should().Be("PT43800H");
+                createdPolicyRule.Type.Should().Be(PolicyType.AccessPolicy);
+            }
+            finally
+            {
+                await client.Applications.DeactivateApplicationAsync(createdApp.Id);
+                await client.Applications.DeleteApplicationAsync(createdApp.Id);
+            }
+        }
+
+        [Fact]
+        public async Task CreateProfileEnrollmentPolicyRule()
+        {
+            var client = TestClient.Create();
+
+            var policy = new Policy()
+            {
+                // Name has a maximum of 50 chars
+                Name = $"dotnet-sdk: ProfileEnrollmentPolicy {Guid.NewGuid()}".Substring(0, 50),
+                Type = PolicyType.ProfileEnrollment,
+                Status = "ACTIVE",
+                Description = "The default policy applies in all situations if no other policy applies.",
+            };
+
+            var createdPolicy = await client.Policies.CreatePolicyAsync(policy);
+            var defaultPolicyRule = await createdPolicy.ListPolicyRules().OfType<IProfileEnrollmentPolicyRule>().FirstOrDefaultAsync();
+            var profileAttributes = new List<IProfileEnrollmentPolicyRuleProfileAttribute>();
+            profileAttributes.Add(new ProfileEnrollmentPolicyRuleProfileAttribute
+            {
+                Name = "email",
+                Label = "Email",
+                Required = true,
+            });
+
+            defaultPolicyRule.Actions = new ProfileEnrollmentPolicyRuleActions
+            {
+                ProfileEnrollment = new ProfileEnrollmentPolicyRuleAction
+                {
+                    Access = "ALLOW",
+                    PreRegistrationInlineHooks = null,
+                    ProfileAttributes = profileAttributes,
+                    UnknownUserAction = "DENY",
+                    TargetGroupIds = null,
+                    ActivationRequirements = new ProfileEnrollmentPolicyRuleActivationRequirement
+                    {
+                        EmailVerification = true,
+                    },
+                },
+            };
+
+            try
+            {
+                var createdPolicyRule = await client.Policies.UpdatePolicyRuleAsync<IProfileEnrollmentPolicyRule>(defaultPolicyRule, createdPolicy.Id, defaultPolicyRule.Id);
+                createdPolicyRule.Should().NotBeNull();
+                createdPolicyRule.Name.Should().Be(defaultPolicyRule.Name);
+                createdPolicyRule.Type.Should().Be(PolicyType.ProfileEnrollment.ToString());
+                createdPolicyRule.Actions.Should().NotBeNull();
+                createdPolicyRule.Actions.ProfileEnrollment.Should().NotBeNull();
+                createdPolicyRule.Actions.ProfileEnrollment.Access.Should().Be("ALLOW");
+                createdPolicyRule.Actions.ProfileEnrollment.PreRegistrationInlineHooks.Should().BeNullOrEmpty();
+                createdPolicyRule.Actions.ProfileEnrollment.UnknownUserAction.Should().Be("DENY");
+                createdPolicyRule.Actions.ProfileEnrollment.TargetGroupIds.Should().BeNullOrEmpty();
+                createdPolicyRule.Actions.ProfileEnrollment.ActivationRequirements.EmailVerification.Should().BeTrue();
+                createdPolicyRule.Actions.ProfileEnrollment.ProfileAttributes.Should().HaveCount(1);
+                createdPolicyRule.Actions.ProfileEnrollment.ProfileAttributes.First().Name.Should().Be("email");
+                createdPolicyRule.Actions.ProfileEnrollment.ProfileAttributes.First().Label.Should().Be("Email");
+                createdPolicyRule.Actions.ProfileEnrollment.ProfileAttributes.First().Required.Should().BeTrue();
+            }
+            finally
+            {
                 await client.Policies.DeletePolicyAsync(createdPolicy.Id);
             }
         }