+++ src/Okta.Sdk.UnitTests/HttpRequestShould.cs
@@ -3,12 +3,11 @@
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 // </copyright>
 
+using System.Net.Http;
 using FluentAssertions;
 using NSubstitute;
 using Okta.Sdk.Internal;
 using Okta.Sdk.UnitTests.Internal;
-using System.Net.Http;
-using System.Threading.Tasks;
 using Xunit;
 
 namespace Okta.Sdk.UnitTests