+++ src/Okta.Sdk.UnitTests/Internal/TestHttpMultipartRequest.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Okta.Sdk.Internal;
+
+namespace Okta.Sdk.UnitTests.Internal
+{
+    public class TestHttpMultipartRequest : MultipartHttpRequest
+    {
+        public IPayloadHandler GetPayloadHandler()
+        {
+            return PayloadHandler;
+        }
+    }
+}