+++ src/Okta.Sdk/MultipartHttpRequest.cs
@@ -1,0 +1,18 @@
+﻿// <copyright file="MultipartHttpRequest.cs" company="Okta, Inc">
+// Copyright (c) 2020 - present Okta, Inc. All rights reserved.
+// Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
+// </copyright>
+
+namespace Okta.Sdk
+{
+    /// <summary>
+    /// This class represents a Multipart request.
+    /// </summary>
+    public class MultipartHttpRequest : HttpRequest
+    {
+        /// <summary>
+        /// Gets or sets the file name.
+        /// </summary>
+        public string FileName { get; set; }
+    }
+}